+++ packages/core/src/devices/laundryWasher.ts
@@ -29,6 +29,7 @@ import { LaundryWasherModeServer } from '@matter/node/behaviors/laundry-washer-m
 import { LaundryWasherControls } from '@matter/types/clusters/laundry-washer-controls';
 import { LaundryWasherMode } from '@matter/types/clusters/laundry-washer-mode';
 import { ModeBase } from '@matter/types/clusters/mode-base';
+import { OnOff } from '@matter/types/clusters/on-off';
 import type { OperationalState } from '@matter/types/clusters/operational-state';
 
 // Matterbridge
@@ -160,7 +161,7 @@ export class MatterbridgeLaundryWasherModeServer extends LaundryWasherModeServer
     const device = this.endpoint.stateOf(MatterbridgeServer);
     device.log.info(`MatterbridgeLaundryWasherModeServer initialized: currentMode is ${this.state.currentMode}`);
     // oxlint-disable-next-line typescript/unbound-method
-    this.reactTo(this.agent.get(MatterbridgeOnOffServer).events.onOff$Changed, this.handleOnOffChange);
+    this.reactTo(this.agent.get(MatterbridgeOnOffServer.with(OnOff.Feature.DeadFrontBehavior)).events.onOff$Changed, this.handleOnOffChange);
   }
 
   // Dead Front OnOff Cluster